+++ README.md
@@ -2,10 +2,61 @@
 
 Plataforma de cores para uso em projetos de interface do FCx Labs.
 
-## Instalation and use
+## Installation
 
 `yarn add fcxlabs-colors`
 
 or
 
 `npm i fcxlabs-colors`
+
+## Usage
+
+### stylesheet:
+
+```js
+import { red } from "fcxlabs-colors";
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: red.shades.light35,
+  },
+});
+```
+
+## Colors reference
+
+### Main
+
+- red
+- green
+- pear
+- gray
+
+### Shades
+
+#### light
+
+- light5
+- light15
+- light25
+- light35
+- light45
+- light55
+- light65
+- light75
+- light85
+- light95
+
+#### dark
+
+- dark5
+- dark15
+- dark25
+- dark35
+- dark45
+- dark55
+- dark65
+- dark75
+- dark85
+- dark95
