+++ examples/index.js
@@ -1,0 +1,3 @@
+var fcxColors = require("../lib");
+
+console.log(fcxColors.red.shades.light15);
