+++ lib/colors/main/gray/shades.d.ts
@@ -1,0 +1,25 @@
+export declare const light: {
+    light5: string;
+    light15: string;
+    light25: string;
+    light35: string;
+    light45: string;
+    light55: string;
+    light65: string;
+    light75: string;
+    light85: string;
+    light95: string;
+};
+export declare const dark: {
+    dark5: string;
+    dark15: string;
+    dark25: string;
+    dark35: string;
+    dark45: string;
+    dark55: string;
+    dark65: string;
+    dark75: string;
+    dark85: string;
+    dark95: string;
+};
+//# sourceMappingURL=shades.d.ts.map
