+++ lib/colors/main/pear/shades.js
@@ -1,0 +1,28 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.dark = exports.light = void 0;
+exports.light = {
+    light5: "#C4DA39",
+    light15: "#CADE4E",
+    light25: "#D1E263",
+    light35: "#D7E678",
+    light45: "#DDEA8D",
+    light55: "#E3EDA1",
+    light65: "#E9F1B6",
+    light75: "#F0F5CB",
+    light85: "#F6F9E0",
+    light95: "#FCFDF5",
+};
+exports.dark = {
+    dark5: "#B7CD2D",
+    dark15: "#A4B828",
+    dark25: "#91A223",
+    dark35: "#7D8C1F",
+    dark45: "#6A771A",
+    dark55: "#576115",
+    dark65: "#444C10",
+    dark75: "#30360C",
+    dark85: "#1D2007",
+    dark95: "#0A0B02",
+};
+//# sourceMappingURL=shades.js.map
