+++ lib/index.d.ts
@@ -1,93 +1,5 @@
-export declare const red: {
-    dark95: string;
-    dark85: string;
-    dark75: string;
-    dark65: string;
-    dark55: string;
-    dark45: string;
-    dark35: string;
-    dark25: string;
-    dark15: string;
-    dark5: string;
-    primary: string;
-    light5: string;
-    light15: string;
-    light25: string;
-    light35: string;
-    light45: string;
-    light55: string;
-    light65: string;
-    light75: string;
-    light85: string;
-    light95: string;
-};
-export declare const green: {
-    dark95: string;
-    dark85: string;
-    dark75: string;
-    dark65: string;
-    dark55: string;
-    dark45: string;
-    dark35: string;
-    dark25: string;
-    dark15: string;
-    dark5: string;
-    primary: string;
-    light5: string;
-    light15: string;
-    light25: string;
-    light35: string;
-    light45: string;
-    light55: string;
-    light65: string;
-    light75: string;
-    light85: string;
-    light95: string;
-};
-export declare const pear: {
-    dark95: string;
-    dark85: string;
-    dark75: string;
-    dark65: string;
-    dark55: string;
-    dark45: string;
-    dark35: string;
-    dark25: string;
-    dark15: string;
-    dark5: string;
-    primary: string;
-    light5: string;
-    light15: string;
-    light25: string;
-    light35: string;
-    light45: string;
-    light55: string;
-    light65: string;
-    light75: string;
-    light85: string;
-    light95: string;
-};
-export declare const gray: {
-    dark95: string;
-    dark85: string;
-    dark75: string;
-    dark65: string;
-    dark55: string;
-    dark45: string;
-    dark35: string;
-    dark25: string;
-    dark15: string;
-    dark5: string;
-    primary: string;
-    light5: string;
-    light15: string;
-    light25: string;
-    light35: string;
-    light45: string;
-    light55: string;
-    light65: string;
-    light75: string;
-    light85: string;
-    light95: string;
-};
+export { red } from "./colors/main/red";
+export { green } from "./colors/main/green";
+export { pear } from "./colors/main/pear";
+export { gray } from "./colors/main/gray";
 //# sourceMappingURL=index.d.ts.map
