+++ src/colors/main/gray/index.ts
@@ -1,0 +1,9 @@
+import { light, dark } from "./shades";
+
+export const gray = {
+  primary: "#53565A",
+  shades: {
+    ...light,
+    ...dark,
+  },
+};
