+++ src/colors/main/gray/shades.ts
@@ -1,0 +1,25 @@
+export const light = {
+  light5: "#5C5E62",
+  light15: "#6D6F73",
+  light25: "#7E8083",
+  light35: "#8F9194",
+  light45: "#A0A2A4",
+  light55: "#B2B3B5",
+  light65: "#C3C4C5",
+  light75: "#D4D5D6",
+  light85: "#E5E6E6",
+  light95: "#F6F7F7",
+};
+
+export const dark = {
+  dark5: "#4F5256",
+  dark15: "#47494D",
+  dark25: "#3E4144",
+  dark35: "#36383B",
+  dark45: "#2E2F32",
+  dark55: "#252728",
+  dark65: "#1D1E1F",
+  dark75: "#151617",
+  dark85: "#0C0D0E",
+  dark95: "#040405",
+};
