+++ src/colors/main/green/index.ts
@@ -1,0 +1,9 @@
+import { light, dark } from "./shades";
+
+export const green = {
+  primary: "#7AC143",
+  shades: {
+    ...light,
+    ...dark,
+  },
+};
