+++ src/colors/main/green/shades.ts
@@ -1,0 +1,25 @@
+export const light = {
+  light5: "#81C44C",
+  light15: "#8ECA5F",
+  light25: "#9BD172",
+  light35: "#A9D785",
+  light45: "#B6DD98",
+  light55: "#C3E3AA",
+  light65: "#D0E9BD",
+  light75: "#DEF0D0",
+  light85: "#EBF6E3",
+  light95: "#F8FCF6",
+};
+
+export const dark = {
+  dark5: "#74B740",
+  dark15: "#68A439",
+  dark25: "#5C9132",
+  dark35: "#4F7D2C",
+  dark45: "#436A25",
+  dark55: "#37571E",
+  dark65: "#2B4417",
+  dark75: "#1F3011",
+  dark85: "#121D0A",
+  dark95: "#060A03",
+};
