+++ src/colors/main/pear/index.ts
@@ -1,0 +1,9 @@
+import { light, dark } from "./shades";
+
+export const pear = {
+  primary: "#C1D82F",
+  shades: {
+    ...light,
+    ...dark,
+  },
+};
