+++ src/colors/main/red/index.ts
@@ -1,0 +1,9 @@
+import { light, dark } from "./shades";
+
+export const red = {
+  primary: "#C8102E",
+  shades: {
+    ...light,
+    ...dark,
+  },
+};
