+++ src/colors/main/red/shades.ts
@@ -1,0 +1,24 @@
+export const light = {
+  light5: "#E8273F",
+  light15: "#EB3E53",
+  light25: "#ED5568",
+  light35: "#EF6B7C",
+  light45: "#F28290",
+  light55: "#F499A4",
+  light65: "#F7B0B8",
+  light75: "#F9C6CD",
+  light85: "#FBDDE1",
+  light95: "#FEF4F5",
+};
+
+export const dark = {
+  dark15: "#AA0E27",
+  dark25: "#960C23",
+  dark35: "#820A1E",
+  dark45: "#6E0919",
+  dark55: "#5A0715",
+  dark65: "#460610",
+  dark75: "#32040C",
+  dark85: "#1E0207",
+  dark95: "#0A0102",
+};
