+++ src/index.ts
@@ -1,95 +1,4 @@
-export const red = {
-  dark95: "#0A0102",
-  dark85: "#1E0207",
-  dark75: "#32040C",
-  dark65: "#460610",
-  dark55: "#5A0715",
-  dark45: "#6E0919",
-  dark35: "#820A1E",
-  dark25: "#960C23",
-  dark15: "#AA0E27",
-  dark5: "#BE0F2C",
-  primary: "#C8102E",
-  light5: "#E8273F",
-  light15: "#EB3E53",
-  light25: "#ED5568",
-  light35: "#EF6B7C",
-  light45: "#F28290",
-  light55: "#F499A4",
-  light65: "#F7B0B8",
-  light75: "#F9C6CD",
-  light85: "#FBDDE1",
-  light95: "#FEF4F5",
-};
-
-export const green = {
-  dark95: "#060A03",
-  dark85: "#121D0A",
-  dark75: "#1F3011",
-  dark65: "#2B4417",
-  dark55: "#37571E",
-  dark45: "#436A25",
-  dark35: "#4F7D2C",
-  dark25: "#5C9132",
-  dark15: "#68A439",
-  dark5: "#74B740",
-  primary: "#7AC143",
-  light5: "#81C44C",
-  light15: "#8ECA5F",
-  light25: "#9BD172",
-  light35: "#A9D785",
-  light45: "#B6DD98",
-  light55: "#C3E3AA",
-  light65: "#D0E9BD",
-  light75: "#DEF0D0",
-  light85: "#EBF6E3",
-  light95: "#F8FCF6",
-};
-
-export const pear = {
-  dark95: "#0A0B02",
-  dark85: "#1D2007",
-  dark75: "#30360C",
-  dark65: "#444C10",
-  dark55: "#576115",
-  dark45: "#6A771A",
-  dark35: "#7D8C1F",
-  dark25: "#91A223",
-  dark15: "#A4B828",
-  dark5: "#B7CD2D",
-  primary: "#C1D82F",
-  light5: "#C4DA39",
-  light15: "#CADE4E",
-  light25: "#D1E263",
-  light35: "#D7E678",
-  light45: "#DDEA8D",
-  light55: "#E3EDA1",
-  light65: "#E9F1B6",
-  light75: "#F0F5CB",
-  light85: "#F6F9E0",
-  light95: "#FCFDF5",
-};
-
-export const gray = {
-  dark95: "#040405",
-  dark85: "#0C0D0E",
-  dark75: "#151617",
-  dark65: "#1D1E1F",
-  dark55: "#252728",
-  dark45: "#2E2F32",
-  dark35: "#36383B",
-  dark25: "#3E4144",
-  dark15: "#47494D",
-  dark5: "#4F5256",
-  primary: "#53565A",
-  light5: "#5C5E62",
-  light15: "#6D6F73",
-  light25: "#7E8083",
-  light35: "#8F9194",
-  light45: "#A0A2A4",
-  light55: "#B2B3B5",
-  light65: "#C3C4C5",
-  light75: "#D4D5D6",
-  light85: "#E5E6E6",
-  light95: "#F6F7F7",
-};
+export { red } from "./colors/main/red";
+export { green } from "./colors/main/green";
+export { pear } from "./colors/main/pear";
+export { gray } from "./colors/main/gray";
